requirements.txt
numpy
scipy
pyaudio
clap-detector

# real tree (serial)
pyserial

# virtual (html) tree
websockets

# plotting
matplotlib